+++ src/sksl/SkSLOperators.h
@@ -22,7 +22,7 @@ public:
 
     // Allow implicit conversion from Token::Kind, since this is just a utility wrapper on top.
     Operator(Token::Kind t) : fKind(t) {
-        SkASSERTF(this->isOperator(), "token-kind %d is not an operator", fKind);
+        SkASSERTF(this->isOperator(), "token-kind %d is not an operator", (int)fKind);
     }
 
     enum class Precedence {
