+++ tools/trace/SkDebugfTracer.cpp
@@ -67,7 +67,7 @@ SkEventTracer::Handle SkDebugfTracer::addTraceEvent(char phase,
     bool open = (phase == TRACE_EVENT_PHASE_COMPLETE);
     if (open) {
         const char* category = this->getCategoryGroupName(categoryEnabledFlag);
-        SkDebugf("[% 2d]%s <%s> %s%s #%d {\n", fIndent.size(), fIndent.c_str(), category, name,
+        SkDebugf("[% 2d]%s <%s> %s%s #%d {\n", (int)fIndent.size(), fIndent.c_str(), category, name,
                  args.c_str(), fCnt);
         fIndent.append(" ");
     } else {
@@ -81,5 +81,5 @@ void SkDebugfTracer::updateTraceEventDuration(const uint8_t* categoryEnabledFlag
                                               const char* name,
                                               SkEventTracer::Handle handle) {
     fIndent.resize(fIndent.size() - 1);
-    SkDebugf("[% 2d]%s } %s\n", fIndent.size(), fIndent.c_str(), name);
+    SkDebugf("[% 2d]%s } %s\n", (int)fIndent.size(), fIndent.c_str(), name);
 }
